+++ Lines/GeometryLineRenderer.cs
@@ -22,16 +22,23 @@ namespace EPPZ.Geometry.Lines
 	{
 
 
-		protected void DrawSegment(Segment segment, Color color)	
-		{ DrawLine(segment.a, segment.b, color); }
-
-		protected void DrawSegment(Segment segment, Color color, bool drawNormals)
+		protected void DrawSegment(Segment segment, Color color, bool drawNormals = false)
 		{
 			DrawLine(segment.a, segment.b, color);
 			if (drawNormals)
 			{
 				Vector2 halfie = segment.a + ((segment.b - segment.a) / 2.0f);
 				DrawLine(halfie, halfie + segment.normal * 0.1f, color);
+			}
+		}
+
+		protected void DrawSegmentWithTransform(Segment segment, Color color, Transform transform_, bool drawNormals = false)
+		{
+			DrawLineWithTransform(segment.a, segment.b, color, transform_);
+			if (drawNormals)
+			{
+				Vector2 halfie = segment.a + ((segment.b - segment.a) / 2.0f);
+				DrawLineWithTransform(halfie, halfie + segment.normal * 0.1f, color, transform_);
 			}
 		}